requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
crispy-bootstrap4==2024.10
Django==5.0.7
django-crispy-forms==2.3
django-embed-video==1.4.10
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
whitenoise==6.7.0
